ml/python-service/requirements.txt
tensorflow>=2.20.0
numpy
flask
flask-cors
python-dotenv
Pillow